requirements.txt
# © Roel Duijsings
pandas== 2.0.2
tqdm== 4.65.0
torch== 2.0.1+cu118
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
Pillow== 9.3.0
torch @ pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118